HttpInterface: Complete rewrite of uploadVideo()
The old code was a total mess and didn't even calculate ranges properly.

Now uses Guzzle and uploads perfectly. However, work remains to be done
with the separate configureVideo() function, which is old and broken!

Currently, the videos are uploaded, but not configured, so they don't
appear on the timeline!