HttpInterface: Revised the new exception system!
We now only throw on the most *serious* HTTP status codes; 404 (Not Found) and
429 (Too Many Requests), instead of throwing on *every* non-2xx HTTP code.

This allows us to handle the fact that the API uses many 4xx codes for things
like "400 Bad Request", which still contain a JSON response for us to read!

The core API request handler in the Instagram class already throws nice
exceptions for such bad API replies, as soon as it sees that the response itself
doesn't have an "ok" status code (isOk()). So we actually DIDN'T need to throw
generic Guzzle exceptions for them inside this HttpInterface handler!