SettingsAdapter: Rewrote bugged config system
The previous system was an unmaintanable mess, and furthermore had a bug in the
ternary operator chaining which meant that it wasn't even picking the correct
setting (because you must use parenthesis to chain multiple ternaries).

Now we use a SINGLE function to read command line options and a SINGLE function
to retrieve the option value with the highest precedence (constructor array is
highest, then command line options, then environment, lastly NULL).

This also simplifies the naming of all options, in a non-backwards compatible
way. Because the past system of different settings names depending on where they
were provided from was VERY messy to deal with from a code perspective.

When migrating to 2.x, users will have to adapt to these new setting names:

SettingsFile:

* Path to store settings in -> settings_path

SettingsMysql:

* Table to use -> db_tablename
* (optional) Re-using a PDO connection object -> pdo
* (if not PDO) Connection username -> db_username
* (if not PDO) Connection password -> db_password
* (if not PDO) Database hostname -> db_host
* (if not PDO) Database name -> db_name

Closes #1039.
Closes #1037.