HttpInterface: Fix direct_share, and code cleanup
The merged direct_share function wasn't properly creating
different request-bodies based on the request. And it was
also a mess the way it tried to guess the share-type based
on various combinations of parameters. A nightmare.

So it's been rewritten to use a clear $shareType parameter
instead.

Furthermore, all sharing functions have been renamed to camelcase
for consistency with all other class functions.

Lastly, there's been tons of other code cleanup, documentation
and bugfixes.

This is almost the end of the HttpInterface work.

Only one unfinished function remains after this!