HttpInterface: Change 404 Not Found handling
It turns out that Instagram uses a "404 Not Found" code to indicate
states like "no such user" when doing certain API lookups.

They SHOULD have used something more appropriate like "204 No Content"
to indicate that their API URL was OK but the "user wasn't found".

But since they've settled on using 404, it means that we cannot assume
that 404 means that the API endpoint was invalid.

Therefore, I'm removing the 404 handling, but keeping it commented out
as a clear warning for future API programmers.