Completely overhauled exception message system
Consistency. Consistency. Consistency.

All exceptions are now proper English sentences and end in punctuation (period).

That's the generally recommended structure for exceptions when programming,
so that they are ready for display to end-users if need be.

And we now automatically add finishing punctuation (period) if the exception
message lacks any punctuation. This is to automatically transform Instagram's
inconsistent exception messages into a consistent result.

Furthermore, ALL internally thrown exceptions now have various internal error
codes, and THAT must ALWAYS be the case going forwards. It's the ONLY way that
library users can identify what an exception is about. Messages are unreliable
and subject to change, but error codes always stay the same.

So again: Always tag internal exceptions with a helpful code!