SettingsFile: Total rewrite for speed and safety
The old class was wasteful and dangerous. It always wrote settings to disk
even if they didn't change since their on-disk value. And it performed the
writes using a direct, truncating overwrite which lead to corruption if the
script was ever interrupted.

I've already done some prior rewrites, but decided to rewrite everything now.

The whole SettingsFile adapter class has been rewritten to modernize
every single aspect of it.

- It now validates the data you're setting, by casting it to a string. No more
  risk of attempting to store an object.
- Data loading from disk is done in a much more efficient and stable way.
- We now only write to disk when the new value differs from the on-disk value.
- All writes are performed atomically for full protection against corruption!

Enjoy.