Refactor configure() for less spaghetti
This is as much as I have energy for now. It's at least much cleaner
than the old code with a ton of parameters.

But the whole "configure()" and "configureVideo()" situation is quite
messy. If someone has the energy to merge them into a single function
which perfectly handles photos (timeline/story), videos (timeline/story)
and albums (with both photos and videos in them), then be my guest.

For now, this is as clean as it gets. I'm done!