HttpInterface: Rename request() to api()
Everything in the file was called "request" these days, which was very
confusing! Therefore, I'm renaming request() to the much more correct and
descriptive api() instead, which indicates that it just performs an API
call and that it's up to the api() caller to handle that any way they want.