Removed old exception leftovers
This old code is no longer necessary. All API calls validate themselves
directly within HttpInterface and throw an exception there if they
aren't okay. So this old code would never even execute if it wasn't Ok!

After this commit, the only remaining place in the library that checks isOk is
in HttpInterface, and that's how it should be! All error handling in 1 place!