Futureproof: uploadAlbum -> uploadTimelineAlbum
Instagram has a tendency to introduce multiple versions of the same feature,
and they may someday add a "story albums" feature. So I am renaming this
function right now so that we DON'T get painted into a corner in the future.

All upload-related functions from now on should all be clearly labeled with
either Timeline or Story, so that we are ready for any future Instagram
features without colliding with our existing function names.

The list of uploading functions is now as follows:

- uploadTimelinePhoto()
- uploadStoryPhoto()
- uploadTimelineVideo()
- uploadStoryVideo()
- uploadTimelineAlbum()