STOP PUTTING DEFAULT VALUES IN RESPONSE OBJECTS
Uppercase commit title to get your attention, everyone.

There is never a need to default anything to an empty string.

There is never a need to default anything to FALSE, because NULL is
equal to FALSE in a "NULL == FALSE" comparison or in a "if ! NULL" comparison,
so "if ($item->getHasAudio())" would be false if has_audio is NULL.

There is never a need to default anything to a number, since NULL
can be used in number comparisons and is treated as zero.

The only thing that "defaults" achieve is to LIE and to make it impossible
to check for NULL aka "nothing provided in server response".

Defaults do absolutely NOTHING when there IS a value (because they get
overwritten with the actual value). And they screw things up when there
ISN'T a reply; by lying as if the field had a value from the server.

So people... stop putting default values in the object fields.

Furthermore, stop writing "@var Item[]|null" in the type annotations. We are
annotating for JsonMapper, not for the library users. The "null" is not
necessary. If the property doesn't exist in the server response, JsonMapper
won't map anything to it. And all properties default to null. So if the server
didn't send us any value, ANY property will remain NULL. Hence "|null"
annotation is useless clutter! ANY property in ANY response can be NULL, and we
should NEVER annotate that, since it serves no purpose at all.