Numeric response property
configure_delay_ms is a number (currently always "6500"), but the server sends
it to us as a string, so let's force it to be a number.

As for "result", it's currently a string which always says "deprecated".
But JsonMapper detects strings automatically, so we don't need to force it
to use a string. It would just interfere if Instagram ever changes the type
of the result parameter to something like an array. Better to let JsonMapper
figure out what kinda JSON the server sent us.