Default to more video upload attempts
Instagram's video server is an overloaded piece of garbage, which drops video
upload chunks when it is overloaded. 10 is a lot of attempts, but it's necessary
for a lot of situations. And we know that the user wants to upload the video, so
retries should be fine. Especially since the retry-algorithm is optimized to
detect failures on a per-chunk basis and restart the whole process quickly
whenever it is bugged.

Until we've figured out how to deal with video upload retries within albums,
it's best that we default it very high.

Related to: #1089.