Rename tons of functions for API consistency
I just discovered that there were tons of getters without a "get"-prefix,
which was inconsistent with the other 90% of the getters that DID have
a proper prefix. Especially since related functions had a prefix.

This follows on from previous rename work in time for 2.x of the API.

All actions are now a consistent, simple word such as "comment()" or "like()".

All getters and setters are now "getX()" or "setX()".

Changes in this commit:

- `autoCompleteUserList` -> `getAutoCompleteUserList`
- `timelineFeed` -> `getTimelineFeed`
- `insights` -> `getInsights`
- `mediaInsights` -> `getMediaInsights`
- `megaphoneLog` -> `getMegaphoneLog`
- `explore` -> `getExplore`
- `discoverChannels` -> `getDiscoverChannels`
- `mediaInfo` -> `getMediaInfo`
- `removeSelftag` -> `removeSelfTag` (fix capitalization for consistency)
- `fbUserSearch` -> `searchFBUsers` (for consistency with `searchUsers`
   and with `searchFBLocation`)
- `userFriendship` -> `getUserFriendship`
- `usersFriendship` -> `getUsersFriendship`
- `deleteCommentsBulk` -> `deleteComments` (renamed for consistency with
   `getUserFriendship`/`getUsersFriendship`; we had to either rename this one,
   or rename the friendship one to "getUserFriendshipBulk", for consistency!)

I've read and re-read the full function list probably 20 times over the
past commits with revisions to function names, so hopefully this the last one!