Devices: Handle the optional Brand separately
The manufacturer and the brand had to be separated into two fields,
since the API calls that need to send manufacturer should only
send manufacturer, whereas the user-agent should always merge
manufacturer and brand (if a brand is available).

Therefore, the Device class now handles manufacturer and brand as
separate fields, and the UserAgent class itself simply builds a
merged string if necessary for *that* particular agent.