<<< BOOM!! PSR-4!! >>>
This is it! The library is now FINALLY compliant with PSR-4 class
structures and autoloaders! Everything is clean, well-organized
and object-oriented! No more mishmash folders. No more crazy names.
No more namespace pollution. Everything has been re-categorized.

This makes the project much less bug-prone, and much easier to work on.

And with that... my 3 weeks of hard work on Instagram-API 2.x are
pretty much over at last...!

All that remains now is to wait a little bit for final public testing.

- SteveJobzniak