Rewrote all upload functions to futureproof them
Closes #1066.

This changes all of the configure* and upload* functions.

They no longer take fixed arguments (which was a huge reason
for spaghetti code in the past). Instead, they take a metadata
array, which we can easily extend in the future with more fields
without ever breaking backwards compatibility again.

Enjoy!