Revert getSeconds change, please think...
The "number_format" change:

1. Made no sense. Do things like that for final formatting, not
hardcoded in a function like getSeconds().

2. If you must edit a function, at least fix its "@return" type
declaration. A number_format is a string return type. One of the biggest
problems of this project in the past was that zero of the functions had
matching PHPdocs for their arguments and return types. It was a big mess
and nobody who read the project could figure out what any functions did.
That sloppiness has to end now with 2.x.