Renamed FunctionException to EndpointException
Initially I chose the name "FunctionException" to try to make it clearer
to newbies. But it just leads to very ugly code for professionals. It's
better that we call it what it really is: An API "function" Endpoint
failure exception. Which is to say... an "EndpointException".

With this change, you can "use InstagramAPI\Exception\EndpointException;"
without feeling bad about polluting your own project with some generically
named "FunctionException". The new name, "EndpointException", is very
unique and guaranteed not to clash with anything in most projects.

Taking care of this now means happier professional users for years to come.

Newbies will simply have to learn that "Endpoint == Function" and figure
it out someday. ;-)