Utils: NEW video detail extractor & new verify
I didn't want to have to depend on ffprobe last time, but it's time to
admit that ffmpeg is not made for extracting video details. Now we
REQUIRE that the user has ffprobe on their system too.

It is a utility which analyzes a video file and outputs JSON describing
it perfectly.

Also made the media format checker reusable, so we can check photo
validity too.

Changes:

- Enhanced video details retrieval to 100% accuracy via FFprobe instead.

- New utility function which validates media resolution and aspect ratio.