SETTINGS: COMPLETE REWRITE FROM THE GROUND UP
Nearly 2000 lines of meticulous new code added! 600 lines of code deleted.

This was a HUGE job which took two days, which is why the release of 2.x
has been delayed a bit. But it was inevitable. The old settings method
was such a piece of garbage that we would have kept suffering forever if
it had been allowed to stay for 2.x.

Over the past month, I made many valiant attempts to rescue the old code
and fixed tons of its bugs, but it was finally time to face facts: It
was one of the worst pieces of garbage in the history of programming.

Even after fixing DOZENS of bugs and making it much more stable, the old
settings system still lacked any sense of design or intelligence. It
locked us into a completely locked-down, rigid format which could NEVER
be extended. And users had to constantly delete their database tables to
"start fresh" whenever we inevitably needed to extend our settings in
any way.

It was severely limiting the ability to grow the project, and to support
new Instagram features in the future. Every time a new database-needing
feature comes out, users would have to delete their databases. Fun, right?!

The settings system was the last remaining old code in this library.

So... it was time to... go into the battlefield once more and to write
something brand new for 2.x. Fresh. From the ground up.

The new settings system is advanced and modular, and layered as follows:

- High-level layer (src/Settings/Factory): Bridges the user's easy
  config array to a StorageHandler and the correct Storage backend. We
  only create ONE of this class and we keep its db connection throughout
  the entire script lifetime, so that we *no longer* wastefully connect
  and disconnect from the storage db every time you switch to a
  different Instagram user. (Those constant disconnects/reconnects were
  yet another braindead detail of the terrible old system.)

- src/Settings/StorageHandler: An intermediary layer which connects to a
  Storage backend and implements higher-level logic using the Storage
  backend's building blocks. This layer adds tons of new features to the
  API, such as the ability to search for, rename and delete users from
  the storage backend (great whenever you're deleting accounts or have
  renamed them on Instagram.com and want to keep your old settings by
  moving them to the new name). And it handles intelligent caching of all
  settings so that there are now as FEW reads/writes from the backend
  storage as possible, for GREAT performance boosts! It also handles all
  cookie translation logic, so that the choice between file-based or
  memory-based cookie jars is no longer hardcoded. ANY storage backend
  can choose whether IT wants a file jar or a memory/database jar!

- src/Settings/StorageInterface: The class interface which all Storage
  backends must implement. Its design is very powerful and yet makes it
  very easy to create YOUR OWN custom settings storage backends! Simply
  implement this interface as described. That is -all- you have to do!

- src/Settings/Storage: Three built-in StorageInterface implementations,
  written from the ground up to give you ultra-reliable File, MySQL and
  Memcached settings backends. You can read them for inspiration if you
  want to write your own backends.

All code is brand new. NONE of the old garbage remains! ;-)

And everything is written with thorough data validation everywhere, so
that there will be *no more* random bugs and errors (which we constantly
encountered in the past with the old system).

The totally new File backend is completely corruption-proof, VERY fast
and VERY reliable. Uses a brand new, strong data format on disk. It is
the default storage backend and is perfect for most people's needs.

The MySQL backend is now bug-free and finally allows us to extend the
database anytime we need to WITHOUT users having to drop their old data!
That should be a huge relief for everyone.

The Memcached backend supports multiple users (instead of just one).

However, I have not tested my Memcached backend since there is no
Memcached support for PHP7 yet, so I wrote that one completely
blindly (but with great care as always).

For those who are overwhelmed and don't know where to begin with all of
this, I also wrote an example in "examples/customSettings.php" showing
the usage as it relates to end-users and your own projects. It is very
simple to configure. You don't even have to provide a single parameter
by default. It will auto-select the File storage backend with a default
path in that case.

Just be aware that this is NOT the old system, so don't assume that all
settings parameters are still the same. They have all been renamed.

Read src/Settings/Factory.php to see what configuration parameters the
new system (and its backends) accepts!

I look forward to seeing what storage backends the community decides to
write, now that it is both easy and fun to implement your own custom
settings backends.

Please remember to contribute back to the project if you come up with
something cool and useful! ;-) As long as you've -correctly- implemented
the StorageInterface specification, they will fit right in and make the
world a better place!

Enjoy! ;-)