Client: Include full server response in exceptions
The exceptions will now always contain the full server response. To get
them, just check if the InstagramException "hasResponse()" and then use
"getResponse()".

Due to this change, it makes no sense for the old "checkStatus" flag
anymore. We will instead always throw an exception if there is a
failure. The remaining code using the old checkStatus flag will be
rewritten after this commit.