Instagram: Migrate autocomplete to exception reply
Instead of the old "disable ALL error checking", this now catches the
appropriate exception and returns the response object via that instead.

I'm not entirely sure we even need this code, because the old error
check disabling (written back in the pre-Guzzle days) was probably just
there to stop the ThrottledException type of server error, but since we
now catch that error separately I don't think we actually need to
preserve that old behavior.

So we can probably remove the whole "catch InstagramException" part. But
it's better that I preserve the current behavior just in case there IS a
reason to still disable error checking.