Instagram: Migrate login to exception reply
This completes the final rewrite for the new exception reply feature!

All exceptions that derive from InstagramException now have a
"hasResponse()" method which, if true, lets you get the whole server
reply object via "getResponse()". It shows exactly what the server said,
and contains all extra details about the exception that was thrown.

As a result, people can retrieve the full details of what the server
said. This is a feature which people have requested a lot!

The exception set we have is perfect and covers every critical error AND
also covers all generic per-endpoint errors via EndpointException.

We will not be adding more exceptions. Instead, you should be catching
the specific ones you want (such as EndpointException), and if you need
to look for specific errors you can check the exception's getMessage or
if you need even more information you should now use the new hasResponse
and getResponse to look at the full server-details of what happened!

Enjoy! :-)