Add example of how to identify endpoint errors
This just expands the endpoint error handling example, to show how to
look for specific errors. This is exactly why we do NOT need any more
exceptions. Because users can already find everything they need to know
via EndpointException. And it's much easier to just catch 1 "generic
endpoint error" exception than to have multiple ones to look for. We
already have 17 exceptions in this project, which is BY FAR enough! ;)