ServerMessageThrower: Fix server message matching
This problem was introduced recently when the "server message prettification"
was extended to make even nicer looking error messages.

The exception-throwing code was looking at the final, "prettified"
user-display message when trying to match the server message against the
list of error reasons. It was trying to find a raw, unprettified server
word like "checkpoint_required" in the prettified message, but at that
point the "$serverMessage" had already been modified to be nice and
user-readable and was actually "Checkpoint required." which of course
did not match. As as result, all such "xxx_yyy" keyword exceptions fell
back to the default "EndpointException" instead of the correct one.

It has been fixed now, by always doing the text-scanning on the
unmodified server-message. And it only prettifies it when it's time to
show the error message to the library user, so that we'll never have
this problem again. :-)

Closes #1216.