SQLite: We now have a SQLite storage backend!
This backend was made possible thanks to the restructuring of PDO
handling, which makes it very easy to write more PDO backends.

SQLite is a very good step-up for people who want to "upgrade" from the
default file-based backend, but don't want to run a big database server.

However, don't be afraid to keep using the "file" backend, since it is
rock-solid and never corrupts any data! So the only real reason to move
to the SQLite backend is if you want your data stored in a single file
instead of multiple files.

The choice is yours! :-)