Instagram: Fix tracking of last_experiments time
I noticed a minor issue that /qe/sync/ was getting called twice. First
on the normal (full) login, and then on the login session resume
immediately after that.

Turns out that it was only storing the last_experiments "save timestamp"
when login sessions were resumed. So that's why it calls it twice, since
it didn't realize that it already got the experiments on the full login.

It has been fixed so that it ALWAYS stores the time anytime we save the
experiments, regardless of who/what calls syncFeatures.