Instagram: Refactoring, grouped story functions
This is the first step of refactoring the ~5000 lines of code in
Instagram.php, since it has become way too messy for us to handle.

It will be done in multiple steps.

Most story functions have been moved now. To $ig->story->...

Other functions will be moved to other namespaces too.