Utils: Make ffmpeg/ffprobe paths user configurable
Closes #1267.

This makes it possible to tell the library where to find ffmpeg and
ffprobe binaries. It's useful for servers that don't have the binaries
in the global PATH. You can put them in any folder and simply tell us
the path. This solves the ffmpeg issues for everyone who cannot install
packages on their servers! :-) Just download compiled binaries and put
them somewhere and give us the path to them. That's all you have to do!