Validate maxAttempts in critical locations
If the user tried sentting a maxAttempts value of 0 (or less), it would
cause lots of code to never run, or some code to only partially run. For
example, uploading a video file, but never running the code for the
configure attempts. Or requesting a video upload URL, but never
uploading any data or calling configure.

Lots of strange and bad things could happen.

So now, I've added validation to all maxAttempts parameters in every
function which EITHER uses the value immediately to do a loop, OR which
does some work and then calls another function which relies on a proper
maxAttempts value.

Users can no longer give a number less than 1 for maxAttempts, and
they'll get warned early (before any work begins) about bad values.