ServerMessageThrower: Improved error extraction
Closes #1291.

It turns out that Instagram sometimes returns a very generic "Sorry,
there was a problem with your request." message string when there's an
error. In that case, they usually have a separate "error_type" variable
describing the actual reason.

So the sentry_block error was thrown as a normal "EndpointException"
with an unknown error reason, which is definitely not what we intended.

This new code instead uses the "error_code" as the server message in
that scenario, so that we properly detect the actual error reason from
the API.

Example server JSON: {"message": "Sorry, there was a problem with your
request.", "status": "fail", "error_type": "sentry_block"}

After this fix, we will now throw the proper exception for that and
other related errors:

Fatal error: Uncaught InstagramAPI\Exception\SentryBlockException:
InstagramAPI\Response\SyncResponse: Sentry block. in ...

This should also eliminate all the tickets we keep getting about "Sorry,
there was a problem...". Since people will now see their actual error.