Rename Model/HdProfilePicUrlInfo to ImageCandidate
I just noticed that the original name made no sense and referred to the
"hd_profile_pic_url_info" field on User objects, even though we also use
it for all normal images everywhere.

So I'm renaming this class to ImageCandidate instead, which is a name
inspired by the "candidates" array of Image_Versions2.