ServerMessageThrower: Optimize regexps
This is a micro-optimization, mostly to fix a bad old habit and set a
better standard for any future regexps.

".*?" (non-greedy) is over 2x faster than ".*" (greedy), when more text
needs to be found after the match... Because the greedy match has to
match all the way to the end of line, then backtracks character by
character to try to find the next text after the match.

Non-greedy matching just ends wildcard-matching instantly when it finds
the next match in the sentence text.