Instagram: Make request indentation consistent
Most text editors indent by 4 spaces when pressing Enter after the
request() creation. This meant that our code had a mishmash of
unindented and indented requests.

It's better to settle on the standard, and it makes the code a bit
clearer to read too.

So now every request's params/extra calls are indented by 4 spaces.