Request: Rename addParams() to addParam()
This is for consistency with addPost(). It never made sense to have
plural in "addParams" and singular in "addPost". Especially since
addParams (plural) just adds a single param.

There was even a bug in our code where we had typed "addParam" by
accident. So it makes more sense to rename it to singular for consistency.