Request/People: Remove weird catch from autocomp
This was a remnant from when I rewrote request handling to remove the
"never throw errors" code. This function had used the old "never throw
any errors" instruction, so when I rewrote it to exceptions I made it
catch all exceptions silently.

But that makes no sense. If there is an error, we'll now throw it properly.

I suspect that the only reason it previously used the "never throw
errors" instruction was because the endpoint is throttled, but we now
catch that error specifically and handle only that error, so it makes no
sense to keep catching all other possible errors too.