Client: Re-use JsonMapper instance for performance
Previously, we always created new instances of JsonMapper every time we
wanted to map an object.

If you're not aware, JsonMapper loads the source code files and analyzes
the comments to find the variable types of all properties.

It then keeps an internal cache of the results, so that it doesn't have
to analyze the same class files over and over again.

Unfortunately, we didn't keep our JsonMapper instance, which meant that
we never had any cache, and always had to inspect all source files every
time we mapped a class.

This rewrite fixes the issue. We now use a single JsonMapper instance so
that its class analysis cache stays in full effect!