Response becomes ResponseInterface & ResponseTrait
This is a common code re-use pattern used in many other libraries.

It allows us to define what interface a server response must implement,
and we provide a standard implementation in ResponseTrait which can be
re-used by all responses.

The benefit is that responses no longer have to be derived from
Response, which means cleaner code where we can create responses that
derive from model objects and inherit all properties, without needing
to manually copy everything and attempting to keep the duplicate objects
in sync with each other (which was very error prone and risky).