Rewrite responses inheriting models
A few of our responses are actually delivered as top-level Model objects
at the root of the response, so those have now been rewritten to a new
hierarchy which shares those models (and their properties) with the
responses. This ensures that no code is duplicated, and that our
responses always stay in sync with the model objects.

Unfortunately JSONMapper needs some rather ugly full paths to the
objects when doing this, but it's the best we can do. So I had to place
some warnings within the model objects that are used in this way.