Request/Media: Avoid extra API query on delete
I said I'd sometimes check the changes and review them. Today was one
of those times. ;-) All changes looked great except this one.

Doing a getInfo() query on every delete() is pretty wasteful since all
of the media info is already available when the user wants to delete it.
They're getting the $mediaId parameter from their Item, so they already
have the object with the other data piece we need!

So I changed this to take a 2nd parameter which can be either a word
such as "PHOTO", or can be the raw value from the item, as follows:

$ig->media->delete($item->getId(), $item->getMediaType());

It's unlikely that we'll ever need more than these 2 params for delete(),
so this is clean enough and saves us from doing pointless API queries.

Take care! :-)