Request/Collection: Rename all funcs to shorter
When a RequestCollection has a very specific purpose, such as "Live" or
"Timeline", or in this calse "Collection", we do not (and should not)
repeat the name of the RequestCollection in the function names. Because
it just leads to lots of extra typing. Compare these old and new names:

Old:

$ig->collection->createCollection()
$ig->collection->deleteCollection()
$ig->collection->getCollections()
$ig->collection->editCollection()
$ig->collection->addSavedMediaToCollection()
$ig->collection->removeMediaFromCollection()

New:

$ig->collection->create()
$ig->collection->delete()
$ig->collection->getList()
$ig->collection->edit()
$ig->collection->addMedia()
$ig->collection->removeMedia()

However, in the cases when we have lots of semi-related functions inside
a collection, we have many situations where we have longer names. But in
this situation, the short names are much better. It's now very clear:
"collection-create, collection-delete, etc", with much less typing. :-)