Request/Collection: Put the list function highest
This is just a small tweak to match the general flow of all other
RequestCollections: Which is info-related functions at the top of the
file, then stuff like create-related and then editing, etc. Each
collection has their own related functions order, and in this case
getting a list of all collections is definitely a high priority action
that should be easy to find at the top.

Alright. This concludes the review of ALL code up to this commit! :-)