Revert "Update composer.json"
This reverts commit 7def57847756d94d0cb005d4a5e3b361066fabea.

We do not need "react/http". It's not a part of this library. None of
the library code uses react/http.

It's only used by one of the examples (realtimeHttp.php), which clearly
says at the absolute top:

"* IMPORTANT!
 * You need https://github.com/reactphp/http in order to run this example.
 * $ composer require react/socket "^0.5"
 * $ composer require react/http "^0.6""

Meaning: If people want to make a project based on the code from the
realtimeHttp example, then *they* should include react/http (AND
react/socket) in their OWN project.

The LIBRARY doesn't need it at all. ;-)