ImageAutoResizer: Safer square-image calculations
This code change is really minor and doesn't change anything in effect,
because whenever we are generating a square "expanded" image, our
"expand"'s final canvas size calculations together with the aspect ratio
of "1" always ends up giving us width=height*1, or height=width/1, so the
values of width and height are already identical and it doesn't matter
which side we pick and didn't matter that we were picking the "shortest"
side. But since expansions support expanding the canvas to any size, it
still makes more sense code-wise to CLEARLY pick the largest side of the
destination canvas as the square size whenever we are expanding. This is
just a code-aesthetic change for clarity.